fix(caplin): Fixes for DataColumnSidecar (#18268)
Based on the spec, index should return as a string from the API

[Contributoor](https://github.com/ethpandaops/contributoor) is not
working for Caplin because
[attestantio](https://github.com/attestantio/go-eth2-client/blob/v0.27.2/api/v1/datacolumnsidecarevent.go)
assumes the index to be a string and panics because it's a integer.

The
[swagger](https://ethereum.github.io/beacon-APIs/#/Debug/getDebugDataColumnSidecars)
also says index should be a string.

Additionally, the correct API path for `data_column_sidecars` is
`/eth/v1/debug/beacon/data_column_sidecars`

I also added `block_root` and `slot` to the SSE events

Based on my testing caplin now mirrors the behavior of lighthouse, the
REST API is matching the swagger/spec, and SSE is mirroring that of
lighthouse and nimbus